refactor: update EmployeeAvatar import paths to shared component for consistency and improved code organization
