--- features/employee/components/employee-avatar.tsx
@@ -1,5 +1,0 @@
-// Re-export from shared location for backward compatibility
-export {
-  EmployeeAvatar,
-  employeeInitials,
-} from '@/components/shared/employee-avatar';
