+++ features/employee/index.ts
@@ -2,7 +2,6 @@ export { EmployeeCard } from './components/employee-card';
 export { CurrentProjectsCard } from './components/current-projects-card';
 export { EmployeeProjectsCell } from './components/employee-projects-cell';
 export { EmployeeStatusBadge } from './components/employee-status-badge';
-export { EmployeeAvatar } from './components/employee-avatar';
 export { EmployeeTable } from './components/employee-table';
 export { EmployeeTreeNode } from './components/employee-tree-node';
 export { EmployeeOverviewTab } from './components/employee-overview-tab';
